test-requirements.txt
vcrpy-unittest
requests-mock
